Step 3 (internal namespace): move util/ → internal/util/
Per DIR-INTERNAL-NAMESPACE-punit step 3. Three main + 2 test source
files under org.javai.punit.util (HashUtils, Lazy, supporting
classes) move under org.javai.punit.internal.util.

Touched:
- git mv 5 Java files; rewrite package declarations.
- Bulk-rewrite 6 import sites across modules + punitexamples.
- Architecture-test files excluded from the sweep (string literals
  owned by step 5).

Architecture-test bookkeeping:
- utilUnderInternal rule's freeze-store: empty — at target state.
- publicMustNotImportInternal rule: refrozen.

Tests green; punitexamples composite build green.

Closes ArchUnit rule: utilUnderInternal.
