--- punit-core/archunit_store/239193ee-09df-4395-aa00-c7cabac03227
@@ -1,2 +1,0 @@
-Class <org.javai.punit.util.HashUtils> does reside in a package 'org.javai.punit.util..' in (HashUtils.java:0)
-Class <org.javai.punit.util.Lazy> does reside in a package 'org.javai.punit.util..' in (Lazy.java:0)
