+++ punit-core/archunit_store/5597aef5-7da4-454d-9b90-c35aa63c63b7
@@ -9,6 +9,8 @@ Field <org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.pacingConfig> has
 Field <org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.spec> has type <org.javai.punit.internal.engine.spec.model.ExecutionSpecification> in (ProbabilisticTestVerdictBuilder.java:0)
 Field <org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.suiteBudget> has type <org.javai.punit.internal.engine.budget.SharedBudgetMonitor> in (ProbabilisticTestVerdictBuilder.java:0)
 Field <org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.tokenMode> has type <org.javai.punit.internal.engine.budget.CostBudgetMonitor$TokenMode> in (ProbabilisticTestVerdictBuilder.java:0)
+Method <org.javai.punit.api.covariate.CovariateDeclaration.computeDeclarationHash()> calls method <org.javai.punit.internal.util.HashUtils.sha256(java.lang.String)> in (CovariateDeclaration.java:127)
+Method <org.javai.punit.api.covariate.CovariateDeclaration.computeDeclarationHash()> calls method <org.javai.punit.internal.util.HashUtils.truncateHash(java.lang.String, int)> in (CovariateDeclaration.java:127)
 Method <org.javai.punit.runtime.BaselineEmitter$1.apply(org.javai.punit.api.spec.TypedSpec)> has return type <org.javai.punit.internal.engine.baseline.BaselineRecord> in (BaselineEmitter.java:0)
 Method <org.javai.punit.runtime.BaselineEmitter.composeRecord(org.javai.punit.api.spec.TypedSpec, org.javai.punit.api.spec.SampleSummary, java.lang.String)> calls constructor <org.javai.punit.internal.engine.baseline.BaselineRecord.<init>(java.lang.String, java.lang.String, java.lang.String, java.lang.String, int, java.time.Instant, java.util.Map, org.javai.punit.api.covariate.CovariateProfile, org.javai.punit.internal.engine.baseline.LatencyIndicator)> in (BaselineEmitter.java:185)
 Method <org.javai.punit.runtime.BaselineEmitter.composeRecord(org.javai.punit.api.spec.TypedSpec, org.javai.punit.api.spec.SampleSummary, java.lang.String)> calls constructor <org.javai.punit.internal.engine.baseline.LatencyIndicator.<init>(org.javai.punit.api.LatencyResult, int, int)> in (BaselineEmitter.java:177)
