+++ punit-core/src/main/java/org/javai/punit/internal/util/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * Utility classes for PUnit.
+ *
+ * <p>This package contains general-purpose utilities used throughout
+ * the PUnit framework.
+ *
+ * <h2>Key Classes</h2>
+ * <ul>
+ *   <li>{@link org.javai.punit.internal.util.Lazy} - Lazily-evaluated, memoized value supplier</li>
+ *   <li>{@link org.javai.punit.internal.util.HashUtils} - SHA-256 hashing and hash truncation</li>
+ * </ul>
+ */
+package org.javai.punit.internal.util;
+
