+++ punit-core/src/test/java/org/javai/punit/internal/engine/baseline/BaselineIntegrityTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.nio.file.Path;
 import java.util.Optional;
 
-import org.javai.punit.util.HashUtils;
+import org.javai.punit.internal.util.HashUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
