+++ punit-core/src/test/java/org/javai/punit/internal/util/HashUtilsTest.java
@@ -1,4 +1,4 @@
-package org.javai.punit.util;
+package org.javai.punit.internal.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
